Move listen_to_google_docs_updates file to warehouse folder
